add tab in UserInfo and FavRoute
